+++ history.md
@@ -1,6 +1,6 @@
 # v1.3.14 - 2018/08/21
 
-* Wired up v3 collection-api
+* Wired up v3 collection-api and track-api
 
 # v1.3.13 - 2018/08/20